Add Tokens and Tests for them

+++ part4/Blogs/controllers/blogs.js
@@ -1,18 +1,26 @@
 const blogsRouter = require('express').Router()
 const { request } = require('../app')
 const Blog = require('../models/blog')
+const jwt = require('jsonwebtoken')
+const { userExtractor } = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1, id: 1 })
   response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body)
+blogsRouter.post('/', userExtractor, async (request, response) => {
+  const user = request.user
+
+  const blogData = { ...request.body, user: user._id }
+  const blog = new Blog(blogData)
 
   try {
     const savedBlog = await blog.save()
-    response.status(201).json(savedBlog)
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+    const populatedBlog = await Blog.findById(savedBlog._id).populate('user', { username: 1, name: 1, id: 1 })
+    response.status(201).json(populatedBlog)
   } catch (error) {
     if (error.name === 'ValidationError') {
       response.status(400).json({ error: error.message })
@@ -38,8 +46,18 @@ blogsRouter.get('/:id', async (request, response) => {
   }
 })
 
-blogsRouter.delete('/:id', async (request, response, next) => {
+blogsRouter.delete('/:id', userExtractor, async (request, response, next) => {
+  const user = request.user
+
   try {
+    const blog = await Blog.findById(request.params.id)
+    if (!blog) {
+      return response.status(404).json({ error: 'blog not found' })
+    }
+    // Only allow delete if user matches
+    if (blog.user && blog.user.toString() !== user._id.toString()) {
+      return response.status(403).json({ error: 'only the creator can delete this blog' })
+    }
     await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
   } catch (exception) {
